requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.7.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0

